+++ frame/ethereum/src/lib.rs
@@ -692,8 +692,8 @@ impl<T: Config> Pallet<T> {
 					true,
 				);
 				if let Some(weight_info) = weight_info {
-					if let Some(proof_size_meter) = weight_info.proof_size_meter {
-						*gas_to_weight.proof_size_mut() = proof_size_meter.usage();
+					if let Some(proof_size_usage) = weight_info.proof_size_usage {
+						*gas_to_weight.proof_size_mut() = proof_size_usage;
 					}
 				}
 				Some(gas_to_weight)
