+++ frame/evm/src/lib.rs
@@ -60,6 +60,7 @@ pub mod benchmarking;
 
 #[cfg(test)]
 mod mock;
+pub mod resource;
 pub mod runner;
 #[cfg(test)]
 mod tests;
@@ -268,8 +269,8 @@ pub mod pallet {
 						true,
 					);
 					if let Some(weight_info) = info.weight_info {
-						if let Some(proof_size_meter) = weight_info.proof_size_meter {
-							*gas_to_weight.proof_size_mut() = proof_size_meter.usage();
+						if let Some(proof_size_usage) = weight_info.proof_size_usage {
+							*gas_to_weight.proof_size_mut() = proof_size_usage;
 						}
 					}
 					Some(gas_to_weight)
@@ -355,8 +356,8 @@ pub mod pallet {
 						true,
 					);
 					if let Some(weight_info) = info.weight_info {
-						if let Some(proof_size_meter) = weight_info.proof_size_meter {
-							*gas_to_weight.proof_size_mut() = proof_size_meter.usage();
+						if let Some(proof_size_usage) = weight_info.proof_size_usage {
+							*gas_to_weight.proof_size_mut() = proof_size_usage;
 						}
 					}
 					Some(gas_to_weight)
@@ -443,8 +444,8 @@ pub mod pallet {
 						true,
 					);
 					if let Some(weight_info) = info.weight_info {
-						if let Some(proof_size_meter) = weight_info.proof_size_meter {
-							*gas_to_weight.proof_size_mut() = proof_size_meter.usage();
+						if let Some(proof_size_usage) = weight_info.proof_size_usage {
+							*gas_to_weight.proof_size_mut() = proof_size_usage;
 						}
 					}
 					Some(gas_to_weight)
