+++ frame/evm/src/tests.rs
@@ -18,8 +18,13 @@
 #![cfg(test)]
 
 use super::*;
-use crate::mock::*;
-
+use crate::{
+	mock::*,
+	resource::{
+		ACCOUNT_BASIC_PROOF_SIZE, ACCOUNT_CODES_METADATA_PROOF_SIZE, ACCOUNT_STORAGE_PROOF_SIZE,
+		IS_EMPTY_CHECK_PROOF_SIZE, WRITE_PROOF_SIZE,
+	},
+};
 use frame_support::{
 	assert_ok,
 	traits::{GenesisBuild, LockIdentifier, LockableCurrency, WithdrawReasons},
@@ -28,10 +33,7 @@ use std::{collections::BTreeMap, str::FromStr};
 
 mod proof_size_test {
 	use super::*;
-	use fp_evm::{
-		CreateInfo, ACCOUNT_BASIC_PROOF_SIZE, ACCOUNT_CODES_METADATA_PROOF_SIZE,
-		ACCOUNT_STORAGE_PROOF_SIZE, IS_EMPTY_CHECK_PROOF_SIZE, WRITE_PROOF_SIZE,
-	};
+	use fp_evm::CreateInfo;
 	use frame_support::traits::StorageInfoTrait;
 	// pragma solidity ^0.8.2;
 	// contract Callee {
@@ -181,7 +183,11 @@ mod proof_size_test {
 			let nonce_increases = ACCOUNT_BASIC_PROOF_SIZE * 2;
 			let expected_proof_size = write_cost + is_empty_check + nonce_increases;
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -238,7 +244,11 @@ mod proof_size_test {
 				+ reading_main_contract_len
 				+ is_empty_check + increase_nonce) as u64;
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -291,7 +301,11 @@ mod proof_size_test {
 				+ reading_main_contract_len
 				+ is_empty_check + increase_nonce) as u64;
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -336,7 +350,11 @@ mod proof_size_test {
 				+ IS_EMPTY_CHECK_PROOF_SIZE
 				+ (ACCOUNT_BASIC_PROOF_SIZE * 2);
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -382,7 +400,11 @@ mod proof_size_test {
 				+ IS_EMPTY_CHECK_PROOF_SIZE
 				+ (ACCOUNT_BASIC_PROOF_SIZE * 2);
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -435,7 +457,11 @@ mod proof_size_test {
 			let expected_proof_size =
 				overhead + (number_balance_reads * ACCOUNT_BASIC_PROOF_SIZE) as u64;
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
 	}
@@ -500,7 +526,11 @@ mod proof_size_test {
 				+ reading_main_contract_len
 				+ is_empty_check + increase_nonce) as u64;
 
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(expected_proof_size, actual_proof_size);
 		});
@@ -583,7 +613,11 @@ mod proof_size_test {
 
 			let ratio = <<Test as Config>::GasLimitPovSizeRatio as Get<u64>>::get();
 			let used_gas = result.used_gas;
-			let actual_proof_size = result.weight_info.expect("weight info").proof_size_usage();
+			let actual_proof_size = result
+				.weight_info
+				.expect("weight info")
+				.proof_size_usage
+				.expect("proof size usage");
 
 			assert_eq!(used_gas.standard, U256::from(21_000));
 			assert_eq!(used_gas.effective, U256::from(actual_proof_size * ratio));
